refactors file structure to remove containers directory and include connect methods in component files
