+++ src/components/Output.jsx
@@ -1,7 +1,16 @@
+import { connect } from 'react-redux';
+
 const Output = (props) => (
     <div>
         <h3>{props.name} is currently {props.activity}.</h3>
     </div>
 );
 
-export default Output;
+const mapStateToProps = (state) => {
+    return {
+        name: state.name,
+        activity: state.activity,
+    }
+}
+
+export default connect(mapStateToProps)(Output);
