--- src/containers/OutputContainer.jsx
@@ -1,12 +1,0 @@
-import Output from '../components/Output';
-
-import { connect } from 'react-redux';
-
-const mapStateToProps = (state) => {
-    return {
-        name: state.name,
-        activity: state.activity,
-    }
-}
-
-export default connect(mapStateToProps)(Output);
